+++ src/Silverback.Integration.Testing/Testing/TestingHelper`1.GetConsumer.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Linq;
 using Silverback.Messaging.Broker;
+using Silverback.Messaging.Configuration;
 
 namespace Silverback.Testing;
 
@@ -23,4 +24,18 @@ public abstract partial class TestingHelper
                 endpoint =>
                     endpoint.RawName == endpointName || endpoint.FriendlyName == endpointName)) ??
         throw new InvalidOperationException($"No consumer found for endpoint '{endpointName}'.");
+
+    /// <summary>
+    ///     Returns the <see cref="ConsumerEndpointConfiguration" /> for the specified endpoint.
+    /// </summary>
+    /// <param name="endpointName">
+    ///     The endpoint name. It could be either the topic/queue name or the friendly name.
+    /// </param>
+    /// <returns>
+    ///     The <see cref="ConsumerEndpointConfiguration" />.
+    /// </returns>
+    protected ConsumerEndpointConfiguration? GetConsumerEndpointConfiguration(string endpointName) =>
+        _consumers?
+            .SelectMany(consumer => consumer.EndpointsConfiguration)
+            .FirstOrDefault(configuration => configuration.RawName == endpointName || configuration.FriendlyName == endpointName);
 }