+++ tests/Silverback.Integration.Tests.E2E/Kafka/BasicFixture.cs
@@ -1,7 +1,6 @@
 // Copyright (c) 2024 Sergio Aquilini
 // This code is licensed under MIT license (see LICENSE file for details)
 
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using FluentAssertions;
@@ -42,7 +41,6 @@ public class BasicFixture : KafkaFixture
                             consumer => consumer
                                 .WithGroupId(DefaultGroupId)
                                 .Consume(endpoint => endpoint.ConsumeFrom(DefaultTopicName))))
-                .AddDelegateSubscriber<IAsyncEnumerable<TestEventTwo>>(_ => Task.CompletedTask)
                 .AddIntegrationSpyAndSubscriber());
 
         IPublisher publisher = Host.ScopedServiceProvider.GetRequiredService<IPublisher>();